[ReconcilePtrCasts] Move ReconcilePtrCasts to a dir
